requirements.txt
flask>=3.0.0
groq
python-dotenv
langchain
langchain-groq